used .clone to enable using the same drone model for both the canvas and the drone telemetry. also fixed issue where drone was not moving by adding it to the this of the canvas
